TileButton: fix accidentally visible label when labelVisible=false
If labelVisible=false, the $buttonLabel should never be shown, even
when a label is set. This is different from a normal Button widget,
which cannot be changed in order not to break existing applications.

Jasmine tests were added to test all combinations of label,
labelVisible, iconId and menus. Tests were also added to ButtonSpec
to assert the current (unexpected) behavior.

393740