DoEntityWrapper: use signed ids in service tunnel
DOs might only be known on the backend server and are sent to the UI
server and then to the browser where they are used in rest calls again.
In this case the id needs to be signed on the backend server as the UI
server has no idea which parts of the DO are ids and which are not.
Therefore, the service tunnel needs to use a signed data object mapper.

413664